Move the initial version check to the constructor.
In a previous change I moved all of the construction logic to the CreateServer hook. Unfortunately, the version check is also there, which is a bad thing - where the address of the CreateServer function has changed due to a new version, the version check would never happen.

I moved the version check to the constructor and removed the usage of exceptions in this case. Instead, we just print to stderr and abort if something has gone wrong.