Mono: Rather than relying on static globals, pass a pointer to the free list to C#, and then pass it back to C++.
This is unfortunate, but required. It seems that when C# interops with C++ code, it calls NWNX_Mono.so as if it had never been loaded before - with a fresh set of static global state. For some reason this does not cascade down into NWNXLib, so we are still able to get globals and use the API.