services: Incorrect usage of std::forward
std::forward has basically one use case: 'to cast a templated function parameter (inside the function) to the value category (lvalue or rvalue) the caller used to pass it. This allows rvalue arguments to be passed on as rvalues, and lvalues to be passed on as lvalues, a scheme called “perfect forwarding.”'

It doesn't apply here and I think confuses intent even if it works, which is to force rvalue references.