NWNXLib/Plugin: Change Info constructor parameters
Given the actual usage of plugins to pass const char* in these parameters, there is no point in std::string&& and even if they didn't, the types are still questionable.

Make it explicit that std::strings are getting constructed one way or the other.  One could argue that these reasonably could be const char* parameters.  I don't know the code base well enough tho.

The use of std::forward here is just not right, even if it works.