Fix: Only changes class level when a verified caster level (resistance/effect creation events included) event fires
Fix 2: No longer sets SetCasterLevelModification no longer sets caster level to a 255 or higher

Other additions: Changed it so the caster level override fires after modification, so that it's actually an override.

Changed hook order so that the Creatures hook happens after the tweak hook. Again so that a set override can override the tweak. (bit strange that was by changing it to earliest.. but okay)